+++ src/com/magento/idea/magento2plugin/actions/generation/generator/PreferenceClassGenerator.java
@@ -2,6 +2,7 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
+
 package com.magento.idea.magento2plugin.actions.generation.generator;
 
 import com.intellij.openapi.project.Project;
@@ -13,26 +14,32 @@ import com.magento.idea.magento2plugin.actions.generation.data.PreferenceFileDat
 import com.magento.idea.magento2plugin.actions.generation.generator.util.DirectoryGenerator;
 import com.magento.idea.magento2plugin.actions.generation.generator.util.FileFromTemplateGenerator;
 import com.magento.idea.magento2plugin.bundles.CommonBundle;
+import com.magento.idea.magento2plugin.bundles.ValidatorBundle;
 import com.magento.idea.magento2plugin.indexes.ModuleIndex;
 import com.magento.idea.magento2plugin.magento.files.PhpPreference;
+import com.magento.idea.magento2plugin.magento.packages.File;
 import com.magento.idea.magento2plugin.util.GetFirstClassOfFile;
 import com.magento.idea.magento2plugin.util.GetPhpClassByFQN;
-import com.magento.idea.magento2plugin.bundles.ValidatorBundle;
-import org.jetbrains.annotations.NotNull;
-import javax.swing.*;
-import com.magento.idea.magento2plugin.magento.packages.File;
 import java.util.Properties;
+import javax.swing.JOptionPane;
+import org.jetbrains.annotations.NotNull;
 
 public class PreferenceClassGenerator extends FileGenerator {
-    private PreferenceFileData preferenceFileData;
-    private Project project;
-    private ValidatorBundle validatorBundle;
-    private CommonBundle commonBundle;
+    private final PreferenceFileData preferenceFileData;
+    private final Project project;
+    private final ValidatorBundle validatorBundle;
+    private final CommonBundle commonBundle;
     private final DirectoryGenerator directoryGenerator;
     private final FileFromTemplateGenerator fileFromTemplateGenerator;
     private final GetFirstClassOfFile getFirstClassOfFile;
 
-    public PreferenceClassGenerator(@NotNull PreferenceFileData preferenceFileData, Project project) {
+    /**
+     * Constructor.
+     */
+    public PreferenceClassGenerator(
+            @NotNull final PreferenceFileData preferenceFileData,
+            final Project project
+    ) {
         super(project);
         this.directoryGenerator = DirectoryGenerator.getInstance();
         this.fileFromTemplateGenerator = FileFromTemplateGenerator.getInstance(project);
@@ -43,16 +50,28 @@ public class PreferenceClassGenerator extends FileGenerator {
         this.commonBundle = new CommonBundle();
     }
 
-    public PsiFile generate(String actionName) {
-        PhpClass pluginClass = GetPhpClassByFQN.getInstance(project).execute(preferenceFileData.getPreferenceFqn());
+
+    @Override
+    public PsiFile generate(final String actionName) {
+        PhpClass pluginClass = GetPhpClassByFQN.getInstance(project).execute(
+                preferenceFileData.getPreferenceFqn()
+        );
 
         if (pluginClass == null) {
             pluginClass = createPluginClass(actionName);
         }
 
         if (pluginClass == null) {
-            String errorMessage = validatorBundle.message("validator.file.cantBeCreated", "Preference Class");
-            JOptionPane.showMessageDialog(null, errorMessage, commonBundle.message("common.error"), JOptionPane.ERROR_MESSAGE);
+            final String errorMessage = validatorBundle.message(
+                    "validator.file.cantBeCreated",
+                    "Preference Class"
+            );
+            JOptionPane.showMessageDialog(
+                    null,
+                    errorMessage,
+                    commonBundle.message("common.error"),
+                    JOptionPane.ERROR_MESSAGE
+            );
 
             return null;
         }
@@ -60,29 +79,41 @@ public class PreferenceClassGenerator extends FileGenerator {
         return pluginClass.getContainingFile();
     }
 
-    private PhpClass createPluginClass(String actionName) {
-        PsiDirectory parentDirectory = ModuleIndex.getInstance(project).getModuleDirectoryByModuleName(getPreferenceModule());
-        String[] pluginDirectories = preferenceFileData.getPreferenceDirectory().split(File.separator);
-        for (String pluginDirectory: pluginDirectories) {
-            parentDirectory = directoryGenerator.findOrCreateSubdirectory(parentDirectory, pluginDirectory);
+    private PhpClass createPluginClass(final String actionName) {
+        PsiDirectory parentDirectory = ModuleIndex.getInstance(project)
+                .getModuleDirectoryByModuleName(getPreferenceModule());
+        final String[] pluginDirectories = preferenceFileData.getPreferenceDirectory()
+                .split(File.separator);
+        for (final String pluginDirectory: pluginDirectories) {
+            parentDirectory = directoryGenerator
+                    .findOrCreateSubdirectory(parentDirectory, pluginDirectory);
         }
 
-        Properties attributes = getAttributes();
-        PsiFile pluginFile = fileFromTemplateGenerator.generate(PhpPreference.getInstance(preferenceFileData.getPreferenceClassName()), attributes, parentDirectory, actionName);
+        final Properties attributes = getAttributes();
+        final PsiFile pluginFile = fileFromTemplateGenerator.generate(
+                PhpPreference.getInstance(preferenceFileData.getPreferenceClassName()),
+                attributes,
+                parentDirectory,
+                actionName
+        );
         if (pluginFile == null) {
             return null;
         }
         return getFirstClassOfFile.execute((PhpFile) pluginFile);
     }
 
-    protected void fillAttributes(Properties attributes) {
-        String preferenceClassName = preferenceFileData.getPreferenceClassName();
+    @Override
+    protected void fillAttributes(final Properties attributes) {
+        final String preferenceClassName = preferenceFileData.getPreferenceClassName();
         attributes.setProperty("NAME", preferenceClassName);
         attributes.setProperty("NAMESPACE", preferenceFileData.getNamespace());
+        if (preferenceFileData.isInterface()) {
+            attributes.setProperty("INTERFACE", "interface");
+        }
         if (!preferenceFileData.isInheritClass()) {
             return;
         }
-        String parentClassName = preferenceFileData.getTargetClass().getName();
+        final String parentClassName = preferenceFileData.getTargetClass().getName();
         if (!parentClassName.equals(preferenceClassName)) {
             attributes.setProperty("USE", preferenceFileData.getTargetClass().getPresentableFQN());
             attributes.setProperty("EXTENDS", parentClassName);